change http controllers pages structure
