+++ src/infra/date-validation/dayjs/dayjs-validator.ts
@@ -2,11 +2,11 @@ import { DateValidator } from '@/domain/coupon/application/support/date-validato
 import { Injectable } from '@nestjs/common'
 import dayjs from 'dayjs'
 import 'dayjs/locale/pt-br'
-dayjs.locale('pt-br')
 
 @Injectable()
 export class DayJsValidator implements DateValidator {
   isExpired(date: Date) {
+    dayjs.locale('pt-br')
     const today = dayjs().endOf('day')
     const expireDate = dayjs(date).endOf('day')
     return expireDate.isBefore(today)
