+++ src/infra/http/controllers/store/activate-product.controller.ts
