+++ src/infra/http/controllers/store/get-product-admin.controller.e2e-spec.ts
