+++ src/infra/http/controllers/store/get-product-admin.controller.ts
@@ -1,7 +1,7 @@
 import { BadRequestException, Controller, Get, Param } from '@nestjs/common'
 import { Roles } from '@/infra/auth/roles.decorator'
 import { GetProductUseCase } from '@/domain/store/application/use-cases/get-product'
-import { ProductDetailsPresenter } from '../presenters/product-details-presenter'
+import { ProductDetailsPresenter } from '../../presenters/product-details-presenter'
 
 @Controller('admin/products/:id')
 export class GetProductAdminController {
