+++ src/infra/http/controllers/store/list-brands.controller.ts
@@ -1,8 +1,8 @@
-import { ListProductsUseCase } from '@/domain/store/application/use-cases/list-products'
+import { ListBrandsUseCase } from '@/domain/store/application/use-cases/list-brands'
 import { ZodValidationPipe } from '@/infra/pipes/zod-validation-pipe'
 import { BadRequestException, Controller, Get, Query } from '@nestjs/common'
 import { z } from 'zod'
-import { ProductPresenter } from '../presenters/product-presenter'
+import { BrandPresenter } from '../../presenters/brand-presenter'
 import { Roles } from '@/infra/auth/roles.decorator'
 
 const pageQueryParamSchema = z
@@ -16,26 +16,25 @@ const queryValidationPipe = new ZodValidationPipe(pageQueryParamSchema)
 
 type PageQueryParamSchema = z.infer<typeof pageQueryParamSchema>
 
-@Controller('/admin/products')
-export class ListProductsAdminController {
-  constructor(private listProducts: ListProductsUseCase) {}
+@Controller('/brands')
+export class ListBrandsController {
+  constructor(private listBrands: ListBrandsUseCase) {}
 
   @Get()
   @Roles(['ADMIN'])
   async handle(@Query('page', queryValidationPipe) page: PageQueryParamSchema) {
-    const result = await this.listProducts.execute({
+    const result = await this.listBrands.execute({
       page,
-      isAdmin: true,
     })
 
     if (result.isLeft()) {
       throw new BadRequestException()
     }
 
-    const products = result.value.products.map((product) =>
-      ProductPresenter.toHTTP(product),
+    const brands = result.value.brands.map((brand) =>
+      BrandPresenter.toHTTP(brand),
     )
 
-    return { products }
+    return { brands }
   }
 }
