+++ src/infra/http/controllers/store/upload-image.controller.e2e-spec.ts
