+++ src/infra/http/controllers/store/upload-image.controller.ts
